+++ R/mm.R
@@ -4,16 +4,17 @@
 #' @param formula A formula specifying an outcome (LHS) and conjoint features (RHS) to describe. All RHS variables should be factors.
 #' @param id An RHS formula specifying a variable holding respondent identifiers, to be used for clustering standard errors.
 #' @param weights An (optional) RHS formula specifying a variable holding survey weights.
+#' @param feature_order An (optional) character vector specifying the names of feature (RHS) variables in the order they should be encoded in the resulting data frame.
 #' @param feature_labels A named list of \dQuote{fancy} feature labels to be used in output. By default, the function looks for a \dQuote{label} attribute on each variable in \code{formula} and uses that for pretty printing. This argument overrides those attributes or otherwise provides fancy labels for this purpose. This should be a list with names equal to variables in \code{formula} and character string values; arguments passed here override variable attributes.
 #' @param level A numeric value indicating the significance level at which to calculate confidence intervals for the MMs (by default 0.95, meaning 95-percent CIs are returned).
 #' @param \dots Ignored.
-#' @details \code{mm} provides descriptive representations of conjoint data as marginal means (MMs), which represent the mean outcome across all appearances of a particular conjoint feature level, averaging across all other features. In forced choice conjoint designs, MMs by definition average 0.5 with values above 0.5 indicating features that increase profile favorability and values below 0.5 indicating features that decrease profile favorability. For continuous outcomes, AMMs can take any value in the full range of the outcome. Plotting functionality is provided in \code{\link{plot.mm}}.
+#' @details \code{mm} provides descriptive representations of conjoint data as marginal means (MMs), which represent the mean outcome across all appearances of a particular conjoint feature level, averaging across all other features. In forced choice conjoint designs, MMs by definition average 0.5 with values above 0.5 indicating features that increase profile favorability and values below 0.5 indicating features that decrease profile favorability. For continuous outcomes, AMMs can take any value in the full range of the outcome. Plotting functionality is provided in \code{\link{plot.cj_mm}}.
 #' @examples
 #' data(hainmueller)
 #' set.seed(12345)
 #' hainmueller$weights <- runif(nrow(hainmueller))
 #' mm(hainmueller, ChosenImmigrant ~ Gender + Education + LanguageSkills, id = ~ CaseID, weights = ~ weights)
-#' @seealso \code{\link{plot.mm}}
+#' @seealso \code{\link{plot.cj_mm}}
 #' @import stats
 #' @importFrom survey svydesign svyby svymean
 #' @export
@@ -22,6 +23,7 @@ function(data,
          formula,
          id,
          weights = NULL,
+         feature_order = NULL,
          feature_labels = NULL,
          level = 0.95,
          ...
@@ -35,6 +37,17 @@ function(data,
     
     # function to produce "fancy" feature labels
     feature_labels <- clean_feature_labels(data = data, RHS = RHS, feature_labels = feature_labels)
+    
+    # process feature_order argument
+    if (!is.null(feature_order)) {
+        if (length(RHS) > length(feature_order)) {
+            warning("'feature_order' appears to be missing values")
+        } else if (length(RHS) < length(feature_order)) {
+            warning("'feature_order' appears to have excess values")
+        }
+    } else {
+        feature_order <- RHS
+    }
     
     # get `id` as character string
     idvar <- all.vars(update(id, 0 ~ . ))
@@ -50,7 +63,7 @@ function(data,
     feature_labels <- clean_feature_labels(data = data, RHS = RHS, feature_labels = feature_labels)
     
     # convert feature labels and levels to data frame
-    term_labels_df <- make_term_labels_df(data, RHS)
+    term_labels_df <- make_term_labels_df(data, feature_order)
     
     # reshape data
     long <- stats::reshape(data[c(outcome, RHS, idvar, weightsvar)], 
@@ -70,21 +83,23 @@ function(data,
     }
     
     # calculate AMMs, SEs, etc.
-    out <- survey::svyby(~ OUTCOME, ~ Level, FUN = survey::svymean, design = svylong)
-    out$z <- out$OUTCOME/out$se
-    out$p <- 2*stats::pnorm(-out$z)
-    out$lower <- out$OUTCOME - stats::qnorm(level + ((1-level)/2)) * out$se
-    out$upper <- out$OUTCOME + stats::qnorm(level + ((1-level)/2)) * out$se
-    names(out) <- c("level", "estimate", "std.error", "z", "p", "lower", "upper")
+    coef_dat <- survey::svyby(~ OUTCOME, ~ Level, FUN = survey::svymean, design = svylong)
+    coef_dat$z <- coef_dat$OUTCOME/coef_dat$se
+    coef_dat$p <- 2*stats::pnorm(-coef_dat$z)
+    coef_dat$lower <- coef_dat$OUTCOME - stats::qnorm(level + ((1-level)/2)) * coef_dat$se
+    coef_dat$upper <- coef_dat$OUTCOME + stats::qnorm(level + ((1-level)/2)) * coef_dat$se
+    names(coef_dat) <- c("level", "estimate", "std.error", "z", "p", "lower", "upper")
     
     # attach feature labels
-    out <- merge(out, make_term_labels_df(data, RHS), by = c("level"), all = TRUE)
-    out$feature <- factor(unlist(unname(feature_labels[match(out$feature, names(feature_labels))])))
-    out$outcome <- outcome
+    coef_dat <- merge(coef_dat, make_term_labels_df(data, RHS), by = c("level"), all = TRUE)
+    coef_dat$feature <- factor(coef_dat$feature,
+                               levels = feature_order,
+                               labels = feature_labels[feature_order])
+    coef_dat$outcome <- outcome
     
     # return organized data frame
-    out <- out[c("outcome", "feature", "level", "estimate", "std.error", "z", "p", "lower", "upper")]
-    out <- out[order(out$level),]
-    rownames(out) <- seq_len(nrow(out))
-    return(structure(out, class = c("mm", "data.frame")))
+    coef_dat <- coef_dat[c("outcome", "feature", "level", "estimate", "std.error", "z", "p", "lower", "upper")]
+    coef_dat <- coef_dat[order(coef_dat$level),]
+    rownames(coef_dat) <- seq_len(nrow(coef_dat))
+    return(structure(coef_dat, class = c("cj_mm", "data.frame")))
 }