+++ NEWS.md
@@ -1,3 +1,10 @@
+# CHANGES TO cregg 0.1.2
+
+* Changed name of `freq()` to `freqs()` and prefixed class names of return values from all functions with `cj_*`.
+* Added `feature_order` argument to all functions to regulate display order.
+* Fixed a bug in the handling of `header_fmt` in `plot()`.
+* Updated README with example of `freqs()`.
+
 # CHANGES TO cregg 0.1.1
 
 * Initial release.